// src/interfaces/Customer.type.ts
import { Contact } from "./Contact.type";
import { CustomerCampaign } from "./CustomerCampaign.type";
import { Purchase } from "./Purchase.type";
import { ReceptiveCall } from "./ReceptiveCall.type";

export type Customer = {
    readonly CODIGO: number;
    RAZAO: string;
    FANTASIA: string;
    PESSOA: 'FIS' | 'JUR';
    ATIVO: 'SIM' | 'NAO';
    CPF_CNPJ: string;
    GRUPO: number;
    END_RUA: string;
    CIDADE: string;
    BAIRRO: string;
    ESTADO: string;
    CEP: string;
    COMPLEMENTO: string;
    ORIGEM: number;
    OPERADOR: number;
    COD_UNIDADE: number;
    AREA1: number;
    FONE1: string | null;
    DESC_FONE1: string;
    COD_ERP: string;
    AREA2: number;
    AREA3: number;
    AREAFAX: number;
    FONE2: string | null;
    FONE3: string | null;
    FAX: string;
    IE_RG: string
    DESC_FONE2: string;
    DESC_FONE3: string;
    DESCFAX: string;
    EMAIL: string | null;
    EMAIL2: string | null;
    WEBSITE: string;
    DATACAD: string;
    STATUS_CAD: string;
    OBS_ADMIN: string | null;
    OBS_OPERADOR: string | null;
    COD_MIDIA: number;
    BLOCK_COMPRAS: 'SIM' | 'NAO';
    SALDO_DISPONIVEL: number;
    SALDO_LIMITE: number;
    POTENCIAL: number;
    DATA_ULT_COMPRA: string | null;
    SEGMENTO: number;
    ULTI_RESULTADO: string;
    DT_AGENDAMENTO: string;
    COD_CAMPANHA: number;
    COD_RESULTADO: number;
    CONTATO_MAIL: string | null;
    ATUALIZADOR: string;
    OPERADOR_LOGIN: string | null;
    OBS_FONE1: string | null;
    OBS_FONE2: string | null;
    OBS_FONE3: string | null;
    NR_FUNCIONARIOS: number | null;
    OBS_CLIENTES: string | null;
    VENCIMENTO_LIMITE_CREDITO: string | null;
    PERIODO_RECOMPRA: number | null;
    DT_ULTIMO_ORCAMENTO_VENDA: string | null;
    POSSUI_ORCAMENTO: 'S' | 'N';
    POSSUI_VENDA: 'S' | 'N';
    CODIGOPRINCIPAL: number | null;
    CONTATOS?: Array<Contact>;
    COMPRAS?: Array<Purchase>;
    CAMPANHAS?: Array<CustomerCampaign>;
    LIGACOES_RECEP?: Array<ReceptiveCall>;
}